app[ARCHI] split hooks and redux
