+++ packages/app/App.tsx
@@ -3,7 +3,7 @@ import "./credentials";
 // React imports
 import React from "react";
 // Redux imports
-import { StoreProvider } from "./src/hooks/store";
+import { StoreProvider } from "./src/Redux/store";
 // Navigation imports
 import Navigator from "./src/Navigator/BottomNavigator";
 // UI imports
