+++ packages/app/src/hooks/useInput.ts
@@ -1,5 +1,4 @@
 import { useState } from "react";
-import { TextInputChangeEventData, NativeSyntheticEvent } from "react-native";
 
 export default function useInput(initialState = "") {
   const [value, setValue] = useState(initialState);
