+++ packages/app/src/screens/accountTab/SignUpView.tsx
@@ -13,10 +13,10 @@ import {
 import Toast from "react-native-easy-toast";
 import { ScrollView } from "react-native-gesture-handler";
 // Custom hooks imports
-import useInput from "../../hooks/useInput";
+import useInput from "../../Hooks/useInput";
 // Redux import
-import { useStore } from "../../hooks/store";
-import { dispatchUserNew } from "../../hooks/dispatchers";
+import { useStore } from "../../Redux/store";
+import { dispatchUserNew } from "../../Redux/dispatchers";
 // API types imports
 import {
   classifyAuthError,
