+++ packages/app/src/screens/cartTab/CartView.tsx
@@ -4,12 +4,12 @@ import { View } from "react-native";
 import { Text } from "react-native-elements";
 import { ButtonValid } from "../../components";
 // Redux import
-import { useStore } from "../../hooks/store";
+import { useStore } from "../../Redux/store";
 // Resources imports
 import { Navigation, Strings } from "../../Resources";
 
 export default function CartScreen({ navigation }) {
-  const { state, dispatch } = useStore();
+  const { state } = useStore();
 
   return (
     <>
